front: Fix comment and and translation
