+++ front/src/common/Map/Search/MapSearchStation.jsx
@@ -6,12 +6,11 @@ import { useTranslation } from 'react-i18next';
 import InputSNCF from 'common/BootstrapSNCF/InputSNCF';
 import { useDebounce } from 'utils/helpers';
 import nextId from 'react-id-generator';
-import turfCenter from '@turf/center';
 import StationCard from 'common/StationCard';
 import { getInfraID } from 'reducers/osrdconf/selectors';
 import { getMap } from 'reducers/map/selectors';
 import { osrdEditoastApi } from 'common/api/osrdEditoastApi';
-import getCoordinates from '../utils';
+import { onResultSearchClick } from '../utils';
 
 export default function MapSearchStation(props) {
   const { updateExtViewport } = props;
@@ -99,27 +98,15 @@ export default function MapSearchStation(props) {
     setSearchResults([...trigramResults, ...nameResults]);
   }, [trigramResults, nameResults]);
 
-  const onResultClick = (result) => {
-    setSearch(result.name);
-
-    const coordinates = getCoordinates(result, map);
-
-    const center = turfCenter(coordinates);
-
-    const newViewport = {
-      ...map.viewport,
-      longitude: center.geometry.coordinates[0],
-      latitude: center.geometry.coordinates[1],
-      zoom: 12,
-    };
-    updateExtViewport(newViewport);
-    dispatch(
-      updateMapSearchMarker({
-        title: result.name,
-        lonlat: [center.geometry.coordinates[0], center.geometry.coordinates[1]],
-      })
-    );
-  };
+  const onResultClick = (result) =>
+    onResultSearchClick({
+      result,
+      map,
+      updateExtViewport,
+      dispatch,
+      title: result.name,
+      setSearch,
+    });
 
   const clearSearchResult = () => {
     setSearch('');
